backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-magic==0.4.27
aioredis==2.0.1
python-json-logger==2.0.7
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
faiss-cpu==1.9.0.post1
sentence-transformers==2.5.1
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.0

# 语音服务依赖
edge-tts>=6.1.17
aiofiles>=23.2.1
funasr>=1.1.12
PyAudio>=0.2.14
webrtcvad>=2.0.10

